Extract generateComponentPath helper
