+++ pkg/combustion/rpm.go
@@ -13,6 +13,7 @@ import (
 )
 
 const (
+	userRPMsDir         = "rpms"
 	modifyRPMScriptName = "10-rpm-install.sh"
 )
 
@@ -20,7 +21,7 @@ const (
 var modifyRPMScript string
 
 func configureRPMs(ctx *image.Context) ([]string, error) {
-	rpmSourceDir, err := generateRPMPath(ctx)
+	rpmSourceDir, err := generateComponentPath(ctx, userRPMsDir)
 	if err != nil {
 		return nil, fmt.Errorf("generating RPM path: %w", err)
 	}
@@ -104,17 +105,4 @@ func writeRPMScript(ctx *image.Context, rpmFileNames []string) (string, error) {
 	}
 
 	return modifyRPMScriptName, nil
-}
-
-func generateRPMPath(ctx *image.Context) (string, error) {
-	rpmSourceDir := filepath.Join(ctx.ImageConfigDir, "rpms")
-	_, err := os.Stat(rpmSourceDir)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return "", nil
-		}
-		return "", fmt.Errorf("checking for RPM directory at %s: %w", rpmSourceDir, err)
-	}
-
-	return rpmSourceDir, nil
 }
